[MBFYCNCT-21] Refactoring
